+++ cardenc/card.hh
@@ -40,7 +40,15 @@ static inline void _encode_atmost(
 	int enc
 )
 {
-	if (enc == enc_cardn)
+	if ((size_t)rhs >= lhs.size())           // nothing to encode
+		return;
+	else if ((size_t)rhs == lhs.size() - 1)  // atmost(n-1) is atleast1
+		common_encode_atmostNm1(dest, lhs);
+	else if (rhs == 0)                       // none of the literals can be true
+		common_encode_atmost0(dest, lhs);
+
+	// general case
+	else if (enc == enc_cardn)
 		cardn_encode_atmostN(top, dest, lhs, rhs);
 	else if (enc == enc_sortn)
 		sortn_encode_atmostN(top, dest, lhs, rhs);
@@ -52,6 +60,8 @@ static inline void _encode_atmost(
 		to_encode_atmostN(top, dest, lhs, rhs);
 	else if (enc == enc_seqc)
 		seqcounter_encode_atmostN(top, dest, lhs, rhs);
+
+	// special case of bitwise, pairwise, and ladder
 	else if (rhs == 1) {
 		if (enc == enc_bitw)
 			bitwise_encode_atmost1(top, dest, lhs);
@@ -72,8 +82,14 @@ static inline void _encode_atleast(
 	int enc
 )
 {
-	if (rhs == 1)
-		common_encode_atleast1(dest, lhs);
+	if (rhs <= 0)                               // nothing to encode
+		return;
+	else if (rhs == 1)
+		common_encode_atleast1(dest, lhs);  // a single clause
+	else if (rhs == lhs.size())
+		common_encode_atleastN(dest, lhs);  // all literals must be true
+
+	// general case
 	else {
 		for (size_t i = 0; i < lhs.size(); ++i)
 			lhs[i] = -lhs[i];